+++ src/xous/syscalls.rs
@@ -9,10 +9,10 @@ use riscv_cpu::cpu::Memory as OtherMemory;
 
 pub fn map_memory(
     memory: &mut Memory,
-    phys: i64,
-    virt: i64,
-    size: i64,
-    _flags: i64,
+    phys: i32,
+    virt: i32,
+    size: i32,
+    _flags: i32,
 ) -> SyscallResult {
     // print!(
     //     "MapMemory(phys: {:08x}, virt: {:08x}, bytes: {}, flags: {:02x})",
@@ -26,8 +26,8 @@ pub fn map_memory(
     }
     if let Some(region) = memory.allocate_virt_region(size as usize) {
         [
-            SyscallResultNumber::MemoryRange as i64,
-            region as i64,
+            SyscallResultNumber::MemoryRange as i32,
+            region as i32,
             size,
             0,
             0,
@@ -43,8 +43,8 @@ pub fn map_memory(
             size as usize
         );
         [
-            SyscallResultNumber::Error as i64,
-            SyscallErrorNumber::OutOfMemory as i64,
+            SyscallResultNumber::Error as i32,
+            SyscallErrorNumber::OutOfMemory as i32,
             0,
             0,
             0,
@@ -65,8 +65,8 @@ pub fn connect(memory: &mut Memory, id: [u32; 4]) -> SyscallResult {
         let connection_id = memory.connections.len() as u32 + 1;
         memory.connections.insert(connection_id, service);
         [
-            SyscallResultNumber::ConnectionId as i64,
-            connection_id as i64,
+            SyscallResultNumber::ConnectionId as i32,
+            connection_id as i32,
             0,
             0,
             0,
@@ -77,7 +77,7 @@ pub fn connect(memory: &mut Memory, id: [u32; 4]) -> SyscallResult {
         .into()
     } else {
         [
-            SyscallResultNumber::ConnectionId as i64,
+            SyscallResultNumber::ConnectionId as i32,
             0,
             0,
             0,
@@ -111,7 +111,7 @@ pub fn send_message(
             *value = memory.read_u8(
                 memory
                     .virt_to_phys(args[0] + offset as u32)
-                    .expect("invalid memory address") as u64,
+                    .expect("invalid memory address"),
             );
         }
         Some(memory_region)
@@ -123,8 +123,8 @@ pub fn send_message(
     let Some(mut service) = memory.connections.remove(&connection_id) else {
         println!("Unhandled connection ID {}", connection_id);
         return [
-            SyscallResultNumber::Error as i64,
-            SyscallErrorNumber::ServerNotFound as i64,
+            SyscallResultNumber::Error as i32,
+            SyscallErrorNumber::ServerNotFound as i32,
             0,
             0,
             0,
@@ -143,12 +143,12 @@ pub fn send_message(
                     services::LendResult::WaitForResponse(msg) => msg.into(),
                     services::LendResult::MemoryReturned(result) => {
                         for (offset, value) in memory_region.into_iter().enumerate() {
-                            memory.write_u8(args[0] as u64 + offset as u64, value);
+                            memory.write_u8(args[0] + offset as u32, value);
                         }
                         [
-                            SyscallResultNumber::MemoryReturned as i64,
-                            result[0] as i64,
-                            result[1] as i64,
+                            SyscallResultNumber::MemoryReturned as i32,
+                            result[0] as i32,
+                            result[1] as i32,
                             0,
                             0,
                             0,
@@ -161,9 +161,9 @@ pub fn send_message(
                 2 => match service.lend(memory, 0, opcode, &memory_region, extra) {
                     services::LendResult::WaitForResponse(msg) => msg.into(),
                     services::LendResult::MemoryReturned(result) => [
-                        SyscallResultNumber::MemoryReturned as i64,
-                        result[0] as i64,
-                        result[1] as i64,
+                        SyscallResultNumber::MemoryReturned as i32,
+                        result[0] as i32,
+                        result[1] as i32,
                         0,
                         0,
                         0,
@@ -174,19 +174,19 @@ pub fn send_message(
                 },
                 3 => {
                     service.send(memory, 0, opcode, &memory_region, extra);
-                    [SyscallResultNumber::Ok as i64, 0, 0, 0, 0, 0, 0, 0].into()
+                    [SyscallResultNumber::Ok as i32, 0, 0, 0, 0, 0, 0, 0].into()
                 }
                 _ => unreachable!(),
             }
         }
         4 => {
             service.scalar(memory, 0, opcode, args);
-            [SyscallResultNumber::Ok as i64, 0, 0, 0, 0, 0, 0, 0].into()
+            [SyscallResultNumber::Ok as i32, 0, 0, 0, 0, 0, 0, 0].into()
         }
         5 => match service.blocking_scalar(memory, 0, opcode, args) {
             services::ScalarResult::Scalar1(result) => [
-                SyscallResultNumber::Scalar1 as i64,
-                result as i64,
+                SyscallResultNumber::Scalar1 as i32,
+                result as i32,
                 0,
                 0,
                 0,
@@ -196,9 +196,9 @@ pub fn send_message(
             ]
             .into(),
             services::ScalarResult::Scalar2(result) => [
-                SyscallResultNumber::Scalar2 as i64,
-                result[0] as i64,
-                result[1] as i64,
+                SyscallResultNumber::Scalar2 as i32,
+                result[0] as i32,
+                result[1] as i32,
                 0,
                 0,
                 0,
@@ -207,12 +207,12 @@ pub fn send_message(
             ]
             .into(),
             services::ScalarResult::Scalar5(result) => [
-                SyscallResultNumber::Scalar5 as i64,
-                result[0] as i64,
-                result[1] as i64,
-                result[2] as i64,
-                result[3] as i64,
-                result[4] as i64,
+                SyscallResultNumber::Scalar5 as i32,
+                result[0] as i32,
+                result[1] as i32,
+                result[2] as i32,
+                result[3] as i32,
+                result[4] as i32,
                 0,
                 0,
             ]
@@ -222,7 +222,7 @@ pub fn send_message(
         _ => {
             panic!("Unknown message kind {}", kind);
             // [
-            //     SyscallResultNumber::Error as i64,
+            //     SyscallResultNumber::Error as i32,
             //     9, /* ServerNotFound */
             //     0,
             //     0,
@@ -248,19 +248,19 @@ pub fn try_send_message(
     send_message(memory, connection_id, kind, opcode, args)
 }
 
-pub fn increase_heap(memory: &mut Memory, delta: i64, _flags: i64) -> SyscallResult {
+pub fn increase_heap(memory: &mut Memory, delta: i32, _flags: i32) -> SyscallResult {
     assert!(delta & 0xfff == 0, "delta must be page-aligned");
     let increase_bytes = delta as u32;
     let heap_address = memory.heap_start + memory.heap_size;
     if delta == 0 {
         return [
-            SyscallResultNumber::MemoryRange as i64,
-            memory.heap_start as i64,
+            SyscallResultNumber::MemoryRange as i32,
+            memory.heap_start as i32,
             if memory.heap_size == 0 {
                 4096
             } else {
                 memory.heap_size
-            } as i64,
+            } as i32,
             0,
             0,
             0,
@@ -271,8 +271,8 @@ pub fn increase_heap(memory: &mut Memory, delta: i64, _flags: i64) -> SyscallRes
     }
     if heap_address.saturating_add(increase_bytes) > super::HEAP_END {
         [
-            SyscallResultNumber::Error as i64,
-            SyscallErrorNumber::OutOfMemory as i64,
+            SyscallResultNumber::Error as i32,
+            SyscallErrorNumber::OutOfMemory as i32,
             0,
             0,
             0,
@@ -288,8 +288,8 @@ pub fn increase_heap(memory: &mut Memory, delta: i64, _flags: i64) -> SyscallRes
         let new_heap_region = memory.heap_start + memory.heap_size;
         memory.heap_size += increase_bytes;
         [
-            SyscallResultNumber::MemoryRange as i64,
-            new_heap_region as i64,
+            SyscallResultNumber::MemoryRange as i32,
+            new_heap_region as i32,
             delta,
             0,
             0,
@@ -303,10 +303,10 @@ pub fn increase_heap(memory: &mut Memory, delta: i64, _flags: i64) -> SyscallRes
 
 pub fn create_thread(
     memory: &mut Memory,
-    entry_point: i64,
-    stack_pointer: i64,
-    stack_length: i64,
-    arguments: [i64; 4],
+    entry_point: i32,
+    stack_pointer: i32,
+    stack_length: i32,
+    arguments: [i32; 4],
 ) -> SyscallResult {
     let (tx, rx) = channel();
     memory
@@ -324,7 +324,7 @@ pub fn create_thread(
         .unwrap();
     let thread_id = rx.recv().unwrap();
     [
-        SyscallResultNumber::ThreadId as i64,
+        SyscallResultNumber::ThreadId as i32,
         thread_id,
         0,
         0,
